Julia Nightly compat, attmpt 2: import is_expected_union
In Julia 1.8 the definition of expected union will change drastically allowing
small unions of concrete types, see discussion in
https://github.com/ulysses4ever/julia-sts/wiki/concrete_type_def-2022-08-15
I'm not sure we want to follow. At the same time, they fix an actual bug in the
old def (where `Union{Missing, Vector}` was considered "expected"). So, for the
time being copy the fix.