Add a filtered selection of "interesting" cases
This adds the columns MethodsDelta and UnstableDelta which contain the difference in the respective value from the last row (e.g. 0 means the value didn't change in the given revision).

These files filter only revisions that have MethodsDelta equal to zero, and UnstableDelta either positive or negative but non-zero. The idea is that if the method count stays the same and at the same time the unstable method count changes, the commit probably made some methods unstable or fixed some unstable ones.

Of course, this is only a heuristic, there is no guarantee that the methods in question are the same (eg. when there is one less unstable method, the commit could've just added a new method and removed the unstable one).

This needs to be investigated by hand by inspecting the commits in question.