Reworked function discovery
This helps detect external method definitions. We now look for functions
in the given module, as well as all of user-provided extra modules, and
then filter all methods of all the discovered functions by whether they
come from the module of interest.

A good option is to pass `Base.loaded_modules_array()` as the extra
modules argument, as this will return a vector of all the modules loaded
currently in the julia session. This includes Base, Core, the module of
interest, and all of its transitive dependencies. Any extra modules that
also happen to be loaded are processed too, but this doesn't seem to
affect performace much.

Document the new `extra_modules` feature